Move Upload funtion into utils and change file.util.tsx to file.util.ts
